requirements.txt
pyautogui
Pillow
pytesseract
keyboard
screeninfo